:bug: consertando bugs no código e nos testes

+++ app/backend/src/domain/usercase/UserUsecase.ts
@@ -3,26 +3,29 @@ import { User } from "../entities/User";
 import { UserRepository } from "../repository/UserRepository";
 
 class UserUseCase {
-  constructor(private userRepository: UserRepository) {}
+  constructor(private userRepository: UserRepository) { }
 
   public registerUser = async (entity: Pick<User, "cpf">): Promise<User> => {
     const userExists = await this.findUserByCPF(entity as Omit<User, "id">)
-    if (userExists) {
+    if (userExists)
       throw new HTTPError(400, 'ExistsCpfException', 'CPF already exists');
-    }
-
-    if (await this.isValidCPF(entity.cpf)) {
+      
+    const isValidCpf = await this.isValidCPF(entity.cpf);
+    if (!isValidCpf)
       throw new HTTPError(404, 'InvalidCpfException', 'CPF is not valid');
-    }
-    
-    if (entity.cpf.length !== 11) {
+
+
+    if (entity.cpf.length !== 11)
       throw new HTTPError(400, 'InvalidCpfException', 'CPF is not valid');
-    }
 
     return await this.userRepository.registerUser(entity);
   }
 
   public findUserByCPF = async (entity: Omit<User, "id">): Promise<Omit<User, 'id'> | null> => {
+    const isValidCpf = await this.isValidCPF(entity.cpf);
+    if (!isValidCpf)
+      throw new HTTPError(404, 'InvalidCpfException', 'CPF is not valid');
+
     const result = await this.userRepository.findUserByCPF(entity);
 
     return result;
@@ -35,32 +38,30 @@ class UserUseCase {
   public removeCPF = async (cpf: string) => {
     const userExists = await this.findUserByCPF({ cpf, createdAt: '' });
 
-    if (!userExists) {
+    if (!userExists)
       throw new HTTPError(400, 'NotFoundCpfException', 'CPF not found');
-    }
+
+    const isValidCpf = await this.isValidCPF(cpf);
+    if (!isValidCpf)
+      throw new HTTPError(404, 'InvalidCpfException', 'CPF is not valid');
 
     return await this.userRepository.removeCPF(cpf);
   }
 
-  public isValidCPF = async (cpf: any)=> {
-    let soma;
-    let resto;
-    soma = 0;
-  if (cpf == "00000000000") return false;
-
-  for (let i = 1; i <= 9; i++) soma = soma + parseInt(cpf.substring(i-1, i)) * (11 - i);
-  resto = (soma * 10) % 11;
-
-    if ((resto == 10) || (resto == 11))  resto = 0;
-    if (resto != parseInt(cpf.substring(9, 10)) ) return false;
-
-  soma = 0;
-    for (let i = 1; i <= 10; i++) soma = soma + parseInt(cpf.substring(i-1, i)) * (12 - i);
-    resto = (soma * 10) % 11;
-
-    if ((resto == 10) || (resto == 11))  resto = 0;
-    if (resto != parseInt(cpf.substring(10, 11) ) ) return false;
-    return true;
+  public isValidCPF = async (cpf: any) => {
+    if (typeof cpf !== 'string') return false
+    cpf = cpf.replace(/[^\d]+/g, '')
+    if (cpf.length !== 11 || !!cpf.match(/(\d)\1{10}/)) return false
+    cpf = cpf.split('')
+    const validator = cpf
+      .filter((digit: any, index: any, array: any) => index >= array.length - 2 && digit)
+      .map((el: any) => +el)
+    const toValidate = (pop: any) => cpf
+      .filter((digit: any, index: any, array: any) => index < array.length - pop && digit)
+      .map((el: any) => +el)
+    const rest = (count: any, pop: any) => (toValidate(pop)
+      .reduce((soma: any, el: any, i: any) => soma + el * (count - i), 0) * 10) % 11 % 10
+    return !(rest(10, 2) !== validator[0] || rest(11, 1) !== validator[1])
   }
 }
